Civil Engineering Insight Studio Project/7.Project Demonstration/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.2.0
